SSL_CERT_FILE, handle redirects
Sets SSL_CERT_FILE to CA Certificate of mitmproxy which allows tools
like curl to work without the "--insecure" option.

Redirects are now handled by resolving the redirect internally during
the capture phase and only storing the final response with the initial
URL, making the redirects invisible to the clients during capture and
replay. This seems to work best with common redirects on, e.g., Github
that only have a short lifetime.